rename things so there are fewer app files
